+++ src/vm/interp/vm_rv64gc.cpp
@@ -4690,9 +4690,9 @@ typename vm_base<ARCH>::virt_addr_t vm_impl<ARCH>::execute_inst(finish_cond_e co
                                     }
                                     else {
                                         uint32_t frs1 = unbox_s(*(F+rs1));
-                                        uint32_t res = fcvt_s(frs1, 0, get_rm((uint8_t)rm));
+                                        int64_t res = (int64_t)(int32_t)fcvt_s(frs1, 0, get_rm(rm));
                                         if((rd) != 0) {
-                                            *(X+rd) = res;
+                                            *(X+rd) = (uint64_t)res;
                                         }
                                         uint32_t flags = fget_flags();
                                         *FCSR = (*FCSR & ~ traits::FFLAG_MASK) | (flags & traits::FFLAG_MASK);
@@ -4724,9 +4724,9 @@ typename vm_base<ARCH>::virt_addr_t vm_impl<ARCH>::execute_inst(finish_cond_e co
                                     }
                                     else {
                                         uint32_t frs1 = unbox_s(*(F+rs1));
-                                        uint32_t res = fcvt_s(frs1, 1, get_rm((uint8_t)rm));
+                                        int64_t res = (int64_t)(int32_t)fcvt_s(frs1, 1, get_rm((uint8_t)rm));
                                         if((rd) != 0) {
-                                            *(X+rd) = res;
+                                            *(X+rd) = (uint64_t)res;
                                         }
                                         uint32_t flags = fget_flags();
                                         *FCSR = (*FCSR & ~ traits::FFLAG_MASK) | (flags & traits::FFLAG_MASK);
@@ -4950,7 +4950,7 @@ typename vm_base<ARCH>::virt_addr_t vm_impl<ARCH>::execute_inst(finish_cond_e co
                                     }
                                     else {
                                         if((rd) != 0) {
-                                            *(X+rd) = (uint64_t)*(F+rs1);
+                                            *(X+rd) = (uint64_t)(int64_t)(int32_t)*(F+rs1);
                                         }
                                     }
                                 }
@@ -5706,7 +5706,7 @@ typename vm_base<ARCH>::virt_addr_t vm_impl<ARCH>::execute_inst(finish_cond_e co
                                         raise(0, traits::RV_CAUSE_ILLEGAL_INSTRUCTION);
                                     }
                                     else {
-                                        uint32_t res = fcvt_64_32(*(F+rs1), 0, get_rm(rm));
+                                        int64_t res = (int64_t)(int32_t)fcvt_64_32(*(F+rs1), 0, get_rm(rm));
                                         if(rd != 0) {
                                             *(X+rd) = (uint64_t)res;
                                         }
@@ -5739,7 +5739,7 @@ typename vm_base<ARCH>::virt_addr_t vm_impl<ARCH>::execute_inst(finish_cond_e co
                                         raise(0, traits::RV_CAUSE_ILLEGAL_INSTRUCTION);
                                     }
                                     else {
-                                        uint32_t res = fcvt_64_32(*(F+rs1), 1, get_rm(rm));
+                                        int64_t res = (int64_t)(int32_t)fcvt_64_32(*(F+rs1), 1, get_rm(rm));
                                         if(rd != 0) {
                                             *(X+rd) = (uint64_t)res;
                                         }
